pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "voltron-evaluation"
authors = [
    {name = "Siddharth Karamcheti", email="skaramcheti@cs.stanford.edu"}
]
description = "Voltron Evaluation: Evaluation Suite for Robotic Representation Learning"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["robotics", "representation learning", "natural language processing", "machine learning"]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "lightning==1.8.6",  # Lock Lightning =>> in case of any breaking future changes...
    "rich",
    "torch",
    "torchvision",
    "torchaudio",
    "voltron-robotics",
]

[project.optional-dependencies]
dev = [
    "black",
    "ipython",
    "pre-commit",
    "ruff",
]

[project.urls]
homepage = "https://github.com/siddk/voltron-evaluation"
repository = "https://github.com/siddk/voltron-evaluation"
documentation = "https://github.com/siddk/voltron-evaluation"

[tool.black]
line-length = 121
target-version = ["py38", "py39", "py310"]
preview = true

[tool.ruff]
line-length = 121
target-version = "py38"
select = ["A", "B", "C90", "E", "F", "I", "RUF", "W"]
ignore = ["B008"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["cache", "data"]
